DjangoAssignment/requirements.txt
asgiref==3.5.2
Django==3.2.8
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mysqlclient==2.1.1
Pillow==8.3.2
PyJWT==2.4.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2